refactor: move duplicate code to util file
